[connector] perf: Speed up deletion crawler
