+++ connectors/src/connectors/notion/temporal/client.ts
@@ -11,15 +11,8 @@ import {
   GARBAGE_COLLECT_QUEUE_NAME,
   QUEUE_NAME,
 } from "@connectors/connectors/notion/temporal/config";
-import type { NotionWebhookEvent } from "@connectors/connectors/notion/temporal/signals";
-import {
-  notionDeletionCrawlSignal,
-  notionWebhookSignal,
-} from "@connectors/connectors/notion/temporal/signals";
-import {
-  notionProcessWebhooksWorkflow,
-  notionSyncWorkflow,
-} from "@connectors/connectors/notion/temporal/workflows/";
+import { notionDeletionCrawlSignal } from "@connectors/connectors/notion/temporal/signals";
+import { notionSyncWorkflow } from "@connectors/connectors/notion/temporal/workflows/";
 import { updateOrphanedResourcesParentsWorkflow } from "@connectors/connectors/notion/temporal/workflows/";
 import { notionDeletionCrawlWorkflow } from "@connectors/connectors/notion/temporal/workflows/deletion_crawl";
 import { notionGarbageCollectionWorkflow } from "@connectors/connectors/notion/temporal/workflows/garbage_collection";
@@ -237,7 +230,7 @@ export async function stopNotionSyncWorkflow(
 
   await stopNotionGarbageCollectorWorkflow(connectorId);
   await stopProcessDatabaseUpsertQueueWorkflow(connectorId);
-  await stopNotionWebhookProcessingWorkflow(connectorId);
+  await stopNotionDeletionCrawlWorkflow(connectorId);
 }
 
 export async function stopNotionGarbageCollectorWorkflow(
@@ -330,20 +323,22 @@ export async function stopProcessDatabaseUpsertQueueWorkflow(
   );
 }
 
-export async function stopNotionWebhookProcessingWorkflow(
+export async function stopNotionDeletionCrawlWorkflow(
   connectorId: ModelId
 ): Promise<void> {
   const connector = await ConnectorResource.fetchById(connectorId);
   if (!connector) {
     throw new Error(`Connector not found. ConnectorId: ${connectorId}`);
   }
-
-  const workflow = await getWebhookProcessingWorkflow(connectorId);
+  const dataSourceConfig = dataSourceConfigFromConnector(connector);
+  const workflow = await getDeletionCrawlWorkflow(connectorId);
 
   if (!workflow) {
-    logger.info(
-      { connectorId },
-      "stopNotionWebhookProcessingWorkflow: Notion webhook processing workflow not found."
+    logger.warn(
+      {
+        workspaceId: dataSourceConfig.workspaceId,
+      },
+      "stopNotionDeletionCrawlWorkflow: Notion deletion crawl workflow not found."
     );
     return;
   }
@@ -351,23 +346,25 @@ export async function stopNotionWebhookProcessingWorkflow(
   const { executionDescription: existingWorkflowExecution, handle } = workflow;
 
   if (existingWorkflowExecution.status.name !== "RUNNING") {
-    logger.info(
-      { connectorId },
-      "stopNotionWebhookProcessingWorkflow: Notion webhook processing workflow is not running."
+    logger.warn(
+      {
+        workspaceId: dataSourceConfig.workspaceId,
+      },
+      "stopNotionDeletionCrawlWorkflow: Notion deletion crawl workflow is not running."
     );
     return;
   }
 
   logger.info(
-    { connectorId },
-    "Terminating existing Notion webhook processing workflow."
+    { workspaceId: dataSourceConfig.workspaceId },
+    "Terminating existing Notion deletion crawl workflow."
   );
 
   await handle.terminate();
 
   logger.info(
-    { connectorId },
-    "Terminated Notion webhook processing workflow."
+    { workspaceId: dataSourceConfig.workspaceId, provider: "notion" },
+    "Terminated Notion deletion crawl workflow."
   );
 }
 
@@ -427,50 +424,6 @@ export async function launchProcessDatabaseUpsertQueueWorkflow(
     { connectorId },
     "launchProcessDatabaseUpsertQueueWorkflow: Started Notion process database upsert queue workflow."
   );
-}
-
-export async function getWebhookProcessingWorkflow(
-  connectorId: ModelId
-): Promise<{
-  executionDescription: WorkflowExecutionDescription;
-  handle: WorkflowHandle;
-} | null> {
-  const client = await getTemporalClient();
-
-  const handle: WorkflowHandle<typeof notionProcessWebhooksWorkflow> =
-    client.workflow.getHandle(
-      getNotionWorkflowId(connectorId, "process-webhooks")
-    );
-
-  try {
-    return { executionDescription: await handle.describe(), handle };
-  } catch (e) {
-    if (e instanceof WorkflowNotFoundError) {
-      return null;
-    }
-    throw e;
-  }
-}
-
-export async function launchNotionWebhookProcessingWorkflow(
-  connectorId: ModelId,
-  event: NotionWebhookEvent
-) {
-  const client = await getTemporalClient();
-
-  await client.workflow.signalWithStart(notionProcessWebhooksWorkflow, {
-    args: [{ connectorId }],
-    taskQueue: QUEUE_NAME,
-    workflowId: getNotionWorkflowId(connectorId, "process-webhooks"),
-    searchAttributes: {
-      connectorId: [connectorId],
-    },
-    signal: notionWebhookSignal,
-    signalArgs: [event],
-    memo: {
-      connectorId,
-    },
-  });
 }
 
 export async function sendDeletionCrawlSignal(
@@ -601,6 +554,27 @@ export async function getProcessDatabaseUpsertQueueWorkflow(
   const handle: WorkflowHandle<typeof processDatabaseUpsertQueueWorkflow> =
     client.workflow.getHandle(
       getNotionWorkflowId(connectorId, "process-database-upsert-queue")
+    );
+
+  try {
+    return { executionDescription: await handle.describe(), handle };
+  } catch (e) {
+    if (e instanceof WorkflowNotFoundError) {
+      return null;
+    }
+    throw e;
+  }
+}
+
+export async function getDeletionCrawlWorkflow(connectorId: ModelId): Promise<{
+  executionDescription: WorkflowExecutionDescription;
+  handle: WorkflowHandle;
+} | null> {
+  const client = await getTemporalClient();
+
+  const handle: WorkflowHandle<typeof notionDeletionCrawlWorkflow> =
+    client.workflow.getHandle(
+      getNotionWorkflowId(connectorId, "deletion-crawl")
     );
 
   try {
